refactor: modularize citizen AI types and patterns
